tweaking pass@k
